Split core type substitution tests
